+++ src/VocabGeneration.test.js
@@ -231,8 +231,12 @@ describe("Suite for generating common vocabularies (marked as [skip] to prevent 
   // it("tests a single custom vocab", async () => {
   it.skip("tests a single custom vocab", async () => {
     await generateVocabArtifact({
-      inputResources: ["https://w3id.org/plasma#"],
-      // inputResources: ["https://raw.githubusercontent.com/coolharsh55/plasma/main/plasma.ttl"],
+      // inputResources: ["https://w3id.org/plasma#"],
+      // inputResources: ["https://w3id.org/oac#"],
+
+      inputResources: [
+        "https://raw.githubusercontent.com/coolharsh55/plasma/main/plasma.ttl",
+      ],
 
       // inputResources: ["http://rdf-vocabulary.ddialliance.org/xkos#"],